# backend/src/test/resources/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 교수업적평가시스템 공통기능 1차 API
  version: 0.1.0
servers:
- url: /api
security:
- sessionCookie: []
...
paths:
...
  /api/roles:
    get:
      operationId: listRoles
      summary: 역할 목록
      x-related-requirements:
      - REQ-015
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      security:
      - sessionCookie: []
      x-roles:
      - R09
      parameters:
      - name: roleCode
        in: query
        required: false
        schema:
          type: string
      - name: roleName
        in: query
        required: false
        schema:
          type: string
      - name: useYn
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/roles/{roleCode}:
    put:
      operationId: updateRole
      summary: 역할 기준 변경
      x-related-requirements:
      - REQ-016
      - REQ-017
      - REQ-005
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      - REQ-062
      security:
      - sessionCookie: []
      x-roles:
      - R09
      x-business-rules:
      - 역할 기준 변경 요청자는 R09 write 권한을 가져야 한다
      - 역할 기준 변경 입력은 Requirement Registry REQ-016, REQ-017, REQ-005의 필수값과 불변 규칙을 만족해야 한다
      x-side-effects:
      - 역할 기준 변경 결과를 관련 PostgreSQL table에 저장한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 추적 가능한 구조에 남긴다
      x-required-tests:
      - 'happy: R09 세션과 유효 입력 -> 역할 기준 변경 후 후속 GET에서 변경값이 보인다'
      - 'auth: 세션 없는 요청 -> 401 ApiError가 반환되고 역할 기준 변경 대상 데이터는 변경되지 않는다'
      - 'validation: 필수 식별자 누락 -> 400 ApiError와 해당 field error가 반환된다'
      - 'business: 역할 기준 변경의 source-backed 불변/기간/권한 규칙 위반 -> 400 또는 409로 차단되고 기존 값이 유지된다'
      - 'side-effect: 역할 기준 변경 성공 -> updated_at과 변경 추적 구조에 처리자/일시가 기록된다'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: roleCode
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
...
  /api/menu-permissions:
    get:
      operationId: listMenuPermissions
      summary: 메뉴 권한 조회
      x-related-requirements:
      - REQ-024
      - REQ-026
      - REQ-027
      - REQ-064
      - REQ-065
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      security:
      - sessionCookie: []
      x-roles:
      - R09
      parameters:
      - name: targetType
        in: query
        required: true
        schema:
          type: string
      - name: targetId
        in: query
        required: true
        schema:
          type: string
      - name: menuId
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    put:
      operationId: saveMenuPermissions
      summary: 메뉴 권한 저장
      x-related-requirements:
      - REQ-025
      - REQ-026
      - REQ-027
      - REQ-064
      - REQ-065
      - REQ-005
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      - REQ-062
      security:
      - sessionCookie: []
      x-roles:
      - R09
      x-business-rules:
      - 메뉴 권한 저장 요청자는 R09 write 권한을 가져야 한다
      - 메뉴 권한 저장 입력은 Requirement Registry REQ-025, REQ-026, REQ-027, REQ-064, REQ-065, REQ-005의 필수값과 불변 규칙을 만족해야 한다
      x-side-effects:
      - 메뉴 권한 저장 결과를 관련 PostgreSQL table에 저장한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 추적 가능한 구조에 남긴다
      x-required-tests:
      - 'happy: R09 세션과 유효 입력 -> 메뉴 권한 저장 후 후속 GET에서 변경값이 보인다'
      - 'auth: 세션 없는 요청 -> 401 ApiError가 반환되고 메뉴 권한 저장 대상 데이터는 변경되지 않는다'
      - 'validation: 필수 식별자 누락 -> 400 ApiError와 해당 field error가 반환된다'
      - 'business: 메뉴 권한 저장의 source-backed 불변/기간/권한 규칙 위반 -> 400 또는 409로 차단되고 기존 값이 유지된다'
      - 'side-effect: 메뉴 권한 저장 성공 -> updated_at과 변경 추적 구조에 처리자/일시가 기록된다'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveMenuPermissionRequest'
...
  /api/menus:
    get:
      operationId: listMenus
      summary: 메뉴 실행정보 조회
      x-related-requirements:
      - REQ-031
      - REQ-034
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      security:
      - sessionCookie: []
      x-roles:
      - R09
      parameters:
      - name: menuName
        in: query
        required: false
        schema:
          type: string
      - name: screenId
        in: query
        required: false
        schema:
          type: string
      - name: url
        in: query
        required: false
        schema:
          type: string
      - name: businessType
        in: query
        required: false
        schema:
          type: string
      - name: useYn
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      operationId: createMenu
      summary: 메뉴 실행정보 등록
      x-related-requirements:
      - REQ-032
      - REQ-034
      - REQ-005
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      - REQ-062
      security:
      - sessionCookie: []
      x-roles:
      - R09
      x-business-rules:
      - 메뉴 실행정보 등록 요청자는 R09 write 권한을 가져야 한다
      - 메뉴 실행정보 등록 입력은 Requirement Registry REQ-032, REQ-034, REQ-005의 필수값과 불변 규칙을 만족해야 한다
      x-side-effects:
      - 메뉴 실행정보 등록 결과를 관련 PostgreSQL table에 저장한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 추적 가능한 구조에 남긴다
      x-required-tests:
      - 'happy: R09 세션과 유효 입력 -> 메뉴 실행정보 등록 후 후속 GET에서 변경값이 보인다'
      - 'auth: 세션 없는 요청 -> 401 ApiError가 반환되고 메뉴 실행정보 등록 대상 데이터는 변경되지 않는다'
      - 'validation: 필수 식별자 누락 -> 400 ApiError와 해당 field error가 반환된다'
      - 'business: 메뉴 실행정보 등록의 source-backed 불변/기간/권한 규칙 위반 -> 400 또는 409로 차단되고 기존 값이 유지된다'
      - 'side-effect: 메뉴 실행정보 등록 성공 -> updated_at과 변경 추적 구조에 처리자/일시가 기록된다'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuRequest'
  /api/menus/{menuId}:
    put:
      operationId: updateMenu
      summary: 메뉴 실행정보 수정
      x-related-requirements:
      - REQ-033
      - REQ-034
      - REQ-005
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      - REQ-062
      security:
      - sessionCookie: []
      x-roles:
      - R09
      x-business-rules:
      - 메뉴 실행정보 수정 요청자는 R09 write 권한을 가져야 한다
      - 메뉴 실행정보 수정 입력은 Requirement Registry REQ-033, REQ-034, REQ-005의 필수값과 불변 규칙을 만족해야 한다
      x-side-effects:
      - 메뉴 실행정보 수정 결과를 관련 PostgreSQL table에 저장한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 추적 가능한 구조에 남긴다
      x-required-tests:
      - 'happy: R09 세션과 유효 입력 -> 메뉴 실행정보 수정 후 후속 GET에서 변경값이 보인다'
      - 'auth: 세션 없는 요청 -> 401 ApiError가 반환되고 메뉴 실행정보 수정 대상 데이터는 변경되지 않는다'
      - 'validation: 필수 식별자 누락 -> 400 ApiError와 해당 field error가 반환된다'
      - 'business: 메뉴 실행정보 수정의 source-backed 불변/기간/권한 규칙 위반 -> 400 또는 409로 차단되고 기존 값이 유지된다'
      - 'side-effect: 메뉴 실행정보 수정 성공 -> updated_at과 변경 추적 구조에 처리자/일시가 기록된다'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: menuId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMenuRequest'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: KNUESESSION
...
  schemas:
    ApiResponse:
      type: object
      required:
      - success
      properties:
        success:
          type: boolean
          enum:
          - true
        data: {}
        message:
          type: string
          nullable: true
    ApiError:
      type: object
      required:
      - success
      - error
      properties:
        success:
          type: boolean
          enum:
          - false
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
            message:
              type: string
            fieldErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
            path:
              type: string
            timestamp:
              type: string
              format: date-time
    RoleCode:
      type: string
      enum:
      - R01
      - R02
      - R03
      - R04
      - R05
      - R06
      - R07
      - R08
      - R09
    UseYn:
      type: string
      enum:
      - Y
      - N
    PermissionTargetType:
      type: string
      enum:
      - ROLE
      - ORGANIZATION
      - USER
    PermissionEffect:
      type: string
      enum:
      - ALLOW
      - DENY
...
    UpdateRoleRequest:
      type: object
      required:
      - roleCode
      - roleName
      - purpose
      - useYn
      properties:
        roleCode:
          $ref: '#/components/schemas/RoleCode'
        roleName:
          type: string
        purpose:
          type: string
        grantCriteria:
          type: string
          nullable: true
        dataScopeDefault:
          type: string
          nullable: true
        useYn:
          $ref: '#/components/schemas/UseYn'
...
    SaveMenuPermissionItem:
      type: object
      required:
      - menuId
      - canRead
      - canWrite
      - effect
      properties:
        menuId:
          type: string
        canRead:
          type: boolean
        canWrite:
          type: boolean
        effect:
          $ref: '#/components/schemas/PermissionEffect'
    SaveMenuPermissionRequest:
      type: object
      required:
      - targetType
      - targetId
      - permissions
      properties:
        targetType:
          $ref: '#/components/schemas/PermissionTargetType'
        targetId:
          type: string
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/SaveMenuPermissionItem'
...
    CreateMenuRequest:
      type: object
      required:
      - menuId
      - menuName
      - displayOrder
      - useYn
      properties:
        menuId:
          type: string
        parentMenuId:
          type: string
          nullable: true
        menuName:
          type: string
        screenId:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        businessType:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        displayOrder:
          type: integer
          minimum: 1
        useYn:
          $ref: '#/components/schemas/UseYn'
    UpdateMenuRequest:
      type: object
      required:
      - menuName
      - displayOrder
      - useYn
      properties:
        parentMenuId:
          type: string
          nullable: true
        menuName:
          type: string
        screenId:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        businessType:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        displayOrder:
          type: integer
          minimum: 1
        useYn:
          $ref: '#/components/schemas/UseYn'
...